Add a simple autonomous Twitch.TV IRC client
 - Add support for a simple "!echo <message>" command via sending a
   PRIVMSG to the channel the client has joined
 - Add primitive support for authenticating the "!echo" command by
   verifying the nick that sent the PRIVMSG

One major notable flaw in this initial implementation is that it does
not attempt to correctly parse messages that are divided across multiple
socket.recv() buffers. Instead, focus was spent on demonstrating how to
parse a wider variety of Twitch.TV IRC commands. No attempts were made
to refactor or simplify many areas of this demonstration code.